fix latent crashes, out-of-bounds clusters and duplicate QC columns
Findings from a review pass over the whole pipeline.

gather_stats.sh summed organelle tag counts with bash integer arithmetic. HOMER
writes those counts as floats; any fractional total (multi-mapper weighting,
bam2td -keepAll) returns from awk in %.6g scientific notation and bash then dies
with "syntax error: invalid arithmetic operator". The sum and rounding now happen
inside awk. The same file carried a comment claiming negative-strand bedGraph
values are stored negative and must be negated: they are positive, and the code
correctly never negated them, so the comment is corrected before someone "fixes"
working code to match it.

collect_consensus_tss.R could write clusters outside their chromosome. Widening
to 150 bp and the overlap-resolution shifts were unclamped with no seqlengths
set, so a TSS near a chromosome end could reach BED with a negative start. Such
clusters are now slid back inside at unchanged width, and one wider than its own
chromosome is left alone and reported. Its symmetric-overlap-pair selection also
relied on the two members of a pair landing adjacent in the Hits object, which
holds for equal-width sorted ranges but is not promised, so it now selects on
queryHits < subjectHits.

tss_size_composition.R built its per-(cluster, length) key in integer arithmetic,
which overflows silently to NA for a large genome with long reads. Now double.

A failed findcsRNATSS.pl call is no longer silent. It still writes an empty TSS
set so one bad library cannot stall a run, but it now warns on stderr and in the
log, since the only other symptom was that library vanishing from the consensus
union plus a confusing bedtools error downstream.

The final QC tables no longer carry FinalRiP, FinalFRiP and FinalEnrichment,
which were literal copies of csRiP, csFRiP and csEnrichment in the same table.
The duplication was hiding a naming subtlety now documented in both the README
and the script: in the final tables those cs* columns are computed against
tss.final.bed while the same names in the initial tables are computed against the
initial merged set, so the two files must not be compared column-by-column.

miRNAPct now shares PretRNAPct's denominator (contaminant over contaminant plus
reads-in-TSS) rather than using nuclear reads. The two are reported side by side
and were on bases differing by roughly the csFRiP factor. Values shift up by
about 15% relative; between-library ratios are essentially unchanged.

Also documented why the second bedtools sort in merge_initial_tss is not
redundant: merge -s emits records in encounter order, interleaving strands, and
the following awk names TSSs by NR, so re-sorting keeps TSS_<n> in coordinate
order.

All four test layers pass; the two fixes the suite does not reach (fractional
organelle counts, out-of-bounds clusters) were verified directly.

+++ workflow/scripts/tss_size_composition.R
@@ -101,18 +101,21 @@ topn <- reads
 
 # Reads in a cluster's top `top_n` lengths, from per-(cluster, length) totals keyed as
 # row + (len - 1) * n_tss. Ranking within a cluster is a sort plus a within-group index,
-# so no per-cluster loop is needed.
+# so no per-cluster loop is needed. The key is built in double rather than integer
+# arithmetic: cluster count times read length exceeds the integer range for a large
+# genome sequenced with long reads (a million clusters and multi-kilobase reads), where
+# integer multiplication would silently return NA.
 top_share <- function(key, val) {
     agg <- rowsum(val, key)
-    k <- as.integer(rownames(agg)); v <- agg[, 1L]
-    row <- ((k - 1L) %% n_tss) + 1L
+    k <- as.numeric(rownames(agg)); v <- agg[, 1L]
+    row <- ((k - 1) %% n_tss) + 1
     o <- order(row, -v)
     row <- row[o]; v <- v[o]
     rank_in_row <- sequence(rle(row)$lengths)
     keep <- rank_in_row <= top_n
     out <- numeric(n_tss)
     agg2 <- rowsum(v[keep], row[keep])
-    out[as.integer(rownames(agg2))] <- agg2[, 1L]
+    out[as.numeric(rownames(agg2))] <- agg2[, 1L]
     out
 }
 
@@ -151,9 +154,9 @@ for (s in seq_along(tagdirs)) {
                 # Collapse to per-(cluster, length) totals now, so what is carried to the
                 # end of the sample stays small.
                 agg <- rowsum(tags$count[sel],
-                    row + (as.integer(tags$len[sel]) - 1L) * n_tss)
+                    row + (as.numeric(tags$len[sel]) - 1) * n_tss)
                 nchunk <- nchunk + 1L
-                len_keys[[nchunk]] <- as.integer(rownames(agg))
+                len_keys[[nchunk]] <- as.numeric(rownames(agg))
                 len_vals[[nchunk]] <- agg[, 1L]
             }
         }
